==> 0. stakeholders_analysis.Rmd <==
---
title: "Stakeholders analysis"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library("alluvial") # alluvial
library(dplyr)
library(ggplot2)
library(ggalluvial)
library(treemap)

di <- "/Users/maria.suarez.munoz/Google Drive/ms_FuzzyCognitiveMapping/FCM_analysis" # Mac

input_folder <- "/input_files/"
output_folder <- "/output_files/"
```

Load file
```{r}
stakeholders <- read.csv(paste(di, input_folder, "interviews_info_clean.csv", sep = ""), sep = ";")

# Fetch table
stakeholders <- stakeholders %>%
  mutate(freq = 1,
         Type = New_stakeholders_classification)
```

Plot
```{r echo = FALSE}
stakeholders %>% group_by(Type, Action.level, Age.code) %>%
  summarise(n = sum(freq)) -> stakeholders3d

ggplot(stakeholders3d, aes(y = n, axis1 = Type, axis2 = Action.level, axis3 = Age.code)) +
  geom_alluvium(col = "lightskyblue3", fill = "lightskyblue3") +
  geom_stratum(fill = "lightskyblue3", color = "grey") +
  geom_text(stat = "stratum", label.strata = TRUE, size = 3) +
  scale_x_continuous(breaks = 1:3, labels = c("Stakeholder type", "Action level", "Age")) +
  scale_y_continuous(NULL, breaks = NULL) +
  ggtitle("Stakeholders' characteristics") +
  theme_minimal()
```

Treemap
```{r}
grouped_st <- stakeholders %>%
  group_by(Type, Knowledge.area...clean) %>%
  select(Type, Knowledge.area...clean) %>%
  summarise(value = n())

treemap(grouped_st,
        index = c("Type","Knowledge.area...clean"),
        vSize = "value",
        type = "index") 
```
