==> 9. matrices_comparisons.Rmd <==
---
title: "Comparison of individual and aggregated matrices"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(FCMapper)
library(rlist)
library(dplyr)

# Directory
di <- "/Users/maria.suarez.munoz/Google Drive/ms_FuzzyCognitiveMapping/FCM_analysis" # Mac

input_folder <- "input_files"
output_folder <- "output_files"
```

# Jaccard

Load individual homogenised matrices
```{r}
files <- list.files(paste(di, input_folder, "homogenised_matrices", sep = "/"), pattern = "*.csv")

# Build maps list
matrices_list <- list()

for (i in 1:length(files)) {
  my_matrix <- read.csv(paste(di, input_folder, "homogenised_matrices", files[i], sep = "/"), sep = ";", na.strings=c("","NA"), dec = ",")
  assign('concept_labels', as.character(my_matrix[,1])) # extract concept names
  my_matrix <- my_matrix[,-1] # discard first column for matrix (concept names)
  colnames(my_matrix) <- concept_labels # to avoid spaces being replaced by dots
  
  matrices_list <- list.append(matrices_list, my_matrix)
}
```

Load aggregated matrix
```{r}
agg_matrix <- read.csv(paste(di, output_folder, "5.1 agg_matrix.csv", sep = "/"), sep = ";", na.strings = c("","NA"), dec = ",")
assign('concept_labels', row.names(agg_matrix)) # extract concept names
colnames(agg_matrix) <- concept_labels # to avoid spaces being replaced by dots
```

Calculate Jaccard index between each individual matrix and the aggregated one
```{r}
jaccard_tb <- data.frame(matrix = integer(),
                            jaccard_with_agg_matrix = numeric())

for (i in 1:length(files)) {
  comparison <- comp.maps(colnames(matrices_list[i][[1]]), colnames(agg_matrix))
  line <- data.frame(matrix = i,
                     jaccard_with_agg_matrix = comparison$Jaccard)
  
  jaccard_tb <- rbind(jaccard_tb, line)
  }
```

# Relationships

Load all relationships table and relationships from aggregated matrix
```{r}
all_relations <- read.csv(paste(di, output_folder, "4.1 all_relations.csv", sep = "/"), sep = ";", na.strings=c("","NA"), dec = ",")

agg_relations_dic <- read.csv(paste(di, output_folder, "4.2 agg_relations_dic.csv", sep = "/"), sep = ";", na.strings=c("","NA"), dec = ",")

nr_relations_in_agg_matrix <- length(agg_relations_dic$source_label)
```

```{r}
unique_relations <- all_relations[,1:3]  %>%
  group_by(matrix, source_label, target_label) %>%
  summarise(count = n())

nr_relationships_per_matrix <- unique_relations %>%
  group_by(matrix) %>%
  summarise(nr_relationships = n()) %>%
  mutate(connections_percentage_with_agg_matrix = (nr_relationships / nr_relations_in_agg_matrix) * 100)
```

```{r}
similarity_tb <- full_join(jaccard_tb, nr_relationships_per_matrix) %>%
  select(matrix, jaccard_with_agg_matrix, connections_percentage_with_agg_matrix)
```

```{r}
write.table(similarity_tb, paste(di, output_folder, "9.1 similarity_homogenised_aggregated.csv", sep = "/"), sep = ";", row.names = FALSE)
```
